Revert "Completed Room DB integration"

+++ app/src/main/java/com/example/moviebox/data/model/movieReview/AuthorDetails.kt
@@ -9,7 +9,7 @@ data class AuthorDetails(
     @SerializedName("name")
     val name: String? = null,
     @SerializedName("rating")
-    val rating: Double? = null,
+    val rating: Int? = null,
     @SerializedName("username")
     val username: String? = null
 )
